+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/src/test/java/com/azure/messaging/eventhubs/checkpointstore/blob/BlobEventProcessorClientStoreTest.java
@@ -275,7 +275,7 @@ public class BlobEventProcessorClientStoreTest {
             any(BlobRequestConditions.class)))
             .thenReturn(Mono.error(new ResourceModifiedException("Etag did not match", null)));
         BlobCheckpointStore blobCheckpointStore = new BlobCheckpointStore(blobContainerAsyncClient);
-        StepVerifier.create(blobCheckpointStore.claimOwnership(Arrays.asList(po))).verifyComplete();
+        StepVerifier.create(blobCheckpointStore.claimOwnership(Arrays.asList(po))).verifyError();
 
         PartitionOwnership po2 = createPartitionOwnership("ns", "eh", "cg", "0", "owner1");
         po2.setETag("1");
@@ -284,11 +284,11 @@ public class BlobEventProcessorClientStoreTest {
         when(blobAsyncClient
             .setMetadataWithResponse(ArgumentMatchers.<Map<String, String>>any(), any(BlobRequestConditions.class)))
             .thenReturn(Mono.error(new ResourceModifiedException("Etag did not match", null)));
-        StepVerifier.create(blobCheckpointStore.claimOwnership(Arrays.asList(po2))).verifyComplete();
+        StepVerifier.create(blobCheckpointStore.claimOwnership(Arrays.asList(po2))).verifyError();
 
         blobCheckpointStore = new BlobCheckpointStore(blobContainerAsyncClient);
         when(blobContainerAsyncClient.getBlobAsyncClient(anyString())).thenReturn(null);
-        StepVerifier.create(blobCheckpointStore.claimOwnership(Arrays.asList(po))).verifyComplete();
+        StepVerifier.create(blobCheckpointStore.claimOwnership(Arrays.asList(po))).verifyError();
     }
 
     private PartitionOwnership createPartitionOwnership(String fullyQualifiedNamespace, String eventHubName,